+++ src/main.cpp
@@ -4,10 +4,10 @@
 int main(int argc, char** argv)
 {
     Eigen::Matrix4f pose = Eigen::Matrix4f::Identity();
-    std::vector<std::pair<float, Eigen::Matrix4f>> poses;
+    std::vector<std::pair<double, Eigen::Matrix4f>> poses;
     for(int i = 0; i < 10; i++)
     {
-        pose(0, 3) = float(i);
+        pose(0, 3) = double(i);
         poses.emplace_back(i, pose);
     }
     Trajectory traj(poses);
